+++ terraform/ces-module/variables.tf
@@ -2,28 +2,21 @@
 variable "component_operator_crd_chart" {
   description = "The helm chart of the component crd. Optional with version like k8s/k8s-component-operator-crd:1.2.3"
   type        = string
-  default     = "k8s/k8s-component-operator-crd:1.10.1"
+  default     = "k8s/k8s-component-operator-crd"
 }
 
 # component operator crd
 variable "blueprint_operator_crd_chart" {
   description = "The helm chart of the blueprint crd. Optional with version like k8s/k8s-blueprint-operator-crd:1.2.3"
   type        = string
-  default     = "k8s/k8s-blueprint-operator-crd:3.1.0"
-}
-
-# component operator image
-variable "component_operator_image" {
-  description = "The Image:Version of the component operator. Optional with version like cloudogu/k8s-component-operator:1.10.0"
-  type        = string
-  default     = "cloudogu/k8s-component-operator:1.10.1"
+  default     = "k8s/k8s-blueprint-operator-crd"
 }
 
 # resource ecosystem itself
 variable "ecosystem_core_chart_version" {
   description = "The version of the ecosystem-core chart"
   type        = string
-  default     = "1.1.0"
+  default     = ""
 }
 
 variable "ecosystem_core_chart_namespace" {